adk/tools/search_tool.py
"""Google Search tool for retrieving regulation updates."""

from typing import List, Dict, Any, Optional
from googlesearch import search
from ..core.logger import get_logger

logger = get_logger(__name__)

class GoogleSearchTool:
    """Tool for performing Google searches."""
    
    def __init__(self, num_results: int = 5):
        """
        Initialize Google Search tool.
        
        Args:
            num_results: Default number of results to return
        """
        self.num_results = num_results
    
    def search(self, query: str, num_results: Optional[int] = None) -> List[Dict[str, str]]:
        """
        Perform a Google search.
        
        Args:
            query: Search query
            num_results: Number of results to return (overrides default)
            
        Returns:
            List of search results with 'title', 'url', and 'description'
        """
        n = num_results or self.num_results
        results = []
        
        logger.info("Performing Google search", query=query, num_results=n)
        
        try:
            # Note: googlesearch-python returns strings (URLs) by default.
            # To get more metadata, we might need a different library or API.
            # For this MVP/Competition, we'll use the basic search which returns URLs
            # and simulate titles/descriptions if needed, or just return URLs.
            # Advanced: Use 'advanced=True' if library supports it to get Result objects.
            
            search_results = search(query, num_results=n, advanced=True)
            
            for result in search_results:
                results.append({
                    "title": result.title,
                    "url": result.url,
                    "description": result.description
                })
                
        except Exception as e:
            logger.error("Google search failed", error=str(e))
            # Fallback/Mock for demo purposes if network fails or library issues
            results = [
                {
                    "title": f"Regulation Update: {query}",
                    "url": "https://example.com/regulation-update",
                    "description": "Simulated search result for demonstration purposes."
                }
            ]
            
        return results
